+++ src/OrdersList/OrdersListSearchConfig.js
@@ -1,17 +1,13 @@
 import moment from 'moment';
 
-import { DATE_FORMAT } from '@folio/stripes-acq-components';
-
 import {
-  CUSTOM_FIELD_TYPES,
-  FILTERS,
-} from './constants';
+  CUSTOM_FIELDS_FILTER,
+  CUSTOM_FIELDS_TYPES,
+  DATE_FORMAT,
+  getCustomFieldsKeywordIndexes,
+} from '@folio/stripes-acq-components';
 
-const indexes = [
-  'metadata.createdDate',
-  'dateOrdered',
-  'poNumber',
-];
+const indexes = ['metadata.createdDate', 'dateOrdered', 'poNumber'];
 
 const searchByDate = (query, dateFormat) => {
   const isoDate = moment.utc(query, dateFormat).format(DATE_FORMAT);
@@ -25,40 +21,33 @@ const customSearchMap = {
 };
 
 function getCqlQuery(query, qindex, dateFormat, customField) {
-  return customField?.type === CUSTOM_FIELD_TYPES.DATE_PICKER
+  return customField?.type === CUSTOM_FIELDS_TYPES.DATE_PICKER
     ? searchByDate(query, dateFormat)
     : customSearchMap[qindex]?.(query, dateFormat) || `*${query}*`;
 }
 
 const getKeywordQuery = (query, dateFormat, customFields) => {
-  const customFieldIndexes = (customFields || [])
-    .filter((cf) => [
-      CUSTOM_FIELD_TYPES.DATE_PICKER,
-      CUSTOM_FIELD_TYPES.TEXTBOX_SHORT,
-      CUSTOM_FIELD_TYPES.TEXTBOX_LONG,
-    ].includes(cf.type))
-    .map((cf) => `${FILTERS.CUSTOM_FIELDS}.${cf.refId}`);
+  const customFieldIndexes = getCustomFieldsKeywordIndexes(customFields);
 
-  return [...indexes, ...customFieldIndexes].reduce(
-    (acc, sIndex) => {
-      const customField = customFields?.find((cf) => `${FILTERS.CUSTOM_FIELDS}.${cf.refId}` === sIndex);
-      const cqlQuery = getCqlQuery(query, sIndex, dateFormat, customField);
+  return [...indexes, ...customFieldIndexes].reduce((acc, sIndex) => {
+    const customField = customFields?.find(
+      (cf) => `${CUSTOM_FIELDS_FILTER}.${cf.refId}` === sIndex,
+    );
+    const cqlQuery = getCqlQuery(query, sIndex, dateFormat, customField);
 
-      if (acc) {
-        return `${acc} or ${sIndex}=="${cqlQuery}"`;
-      } else {
-        return `${sIndex}=="${cqlQuery}"`;
-      }
-    },
-    '',
-  );
+    if (acc) {
+      return `${acc} or ${sIndex}=="${cqlQuery}"`;
+    } else {
+      return `${sIndex}=="${cqlQuery}"`;
+    }
+  }, '');
 };
 
 export function makeSearchQuery(dateFormat, customFields) {
   return (query, qindex) => {
     if (qindex) {
       const customField = customFields?.find(
-        (cf) => `${FILTERS.CUSTOM_FIELDS}.${cf.refId}` === qindex,
+        (cf) => `${CUSTOM_FIELDS_FILTER}.${cf.refId}` === qindex,
       );
 
       const cqlQuery = getCqlQuery(query, qindex, dateFormat, customField);