+++ src/common/index.js
@@ -1,6 +1,5 @@
 export { default as AddressFilter } from './AddressFilter';
 export { default as ClosingReasonFilter } from './ClosingReasonFilter';
-export { default as CustomFieldsFilters } from './CustomFieldsFilters';
 export { ExportDetailsAccordion } from './ExportDetailsAccordion';
 export { ExportDetailsList } from './ExportDetailsList';
 export { ReexportActionButton } from './ReexportActionButton';